requirements.txt
automium==0.2.6
automium-web==0.1.1
Click==7.0
Django==2.2.7
django-model-utils==3.2.0
djangorestframework==3.10.3
ez-setup==0.9
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
legacy==0.1.6
MarkupSafe==1.1.1
PyMySQL==0.9.3
pytz==2019.3
sqlparse==0.3.0
Werkzeug==0.16.0
